Add support for DME and fix IDs
Add support for data module instances using the data module extended
(DME) code for identification. If only one instance of a data module is
included in a publication, its ID is derived only from its data module
code and can therefore be referenced without specifying the extension in
a dmRef. If multiple instances of a data module are included, each
instance's ID is prefixed by the extensionProducer and extensionCode,
and must be referenced using the extension.

All IDs are now prefixed with the string 'ID_', since data module codes
may begin with a numeric character but XML IDs may not.