Support alternative publication format
This format extends the S1000D pub module schema to include data module
elements directly in the pub module. Another tool is used to "flatten"
the pub module containing dmRefs to this format. This should be slightly
more efficient since it allows the order of the data modules in the
publication to be determined by their order in the input, and data
modules can be processed as they appear, rather than iterating over all
data modules in a publication to find the right one.

The original publication format should not be affected, so whichever
suits the situation better can be used.