Add workaround for change marks on list items
Like table and definitionList, randomList, attentionRandomList,
sequentialList and attentionSequentialList have issues with change marks
on their child elements.

The method of adding change marks in the dbtofo stylesheets involves
wrapping elements in an fo:block, but this cannot be done for
fo:list-item.

The workaround is to add a change mark to the entire list if any of its
descendants have a mark.